project/requirements.txt
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
fonttools==4.58.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
shiboken6==6.9.0
six==1.17.0
